+++ src/components/MapSvgLayerPainter.vue
@@ -1,12 +1,12 @@
 <template>
-  <MapSvgLayer id="regions">
+  <MapSvgLayer id="painter">
     <path
       class="region"
       v-for="region in regions"
       :key="region.key"
       :d="region.d"
       :style="style(region)"
-      @click="onClick(region)"
+      @click="paintRegion(region)"
       v-tooltip="tooltipRegionOwner(region)"
     />
   </MapSvgLayer>
@@ -14,55 +14,67 @@
 
 <script>
 import MapSvgLayer from "@/components/MapSvgLayer";
+import { watch, onBeforeUnmount } from '@vue/composition-api';
+import { usePlanner } from "@/use/planner";
 
 export default {
   components: { MapSvgLayer },
-  props: {
-    regions: Object,
-    factions: Object,
-    starting_regions: Object
-  },
-  created() {
-    // this is a reset
-    this.$store.commit("SET_PAINTER_FACTIONS", Object.assign({}, this.starting_regions));
-  },
-  computed: {
-    selected() {
-      return this.$store.state.painter.faction;
-    },
-    factionList() {
-      return this.$store.state.painter.factions;
-    }
-  },
-  methods: {
-    onClick(region) {
-      if (this.selected) {
-        this.$store.commit("SET_PAINTER_FACTIONS_ITEM", {
-          regionKey: region.key,
-          factionKey: this.selected.key
-        });
+  setup() {
+    const {
+      factions,
+      selectedFaction,
+      ownedRegions,
+      regions,
+      startingRegions,
+
+      getStateFromQueryParams,
+      reset
+    } = usePlanner();
+
+    watch(() => startingRegions.value, () => {
+      const fromQuery = getStateFromQueryParams(startingRegions.value, factions);
+      if (fromQuery) {
+        ownedRegions.value = fromQuery;
       } else {
-        this.$store.commit("SET_PAINTER_FACTIONS_ITEM", {
-          regionKey: region.key,
-          factionKey: null
-        });
+        ownedRegions.value = Object.assign({}, startingRegions.value);
       }
-    },
-    style(region) {
-      const owner = this.factionList[region.key];
-      const faction = this.factions[owner];
+    });
+
+    onBeforeUnmount(() => {
+      reset();
+    });
+
+    const style = (region) => {
+      const owner = ownedRegions.value[region.key];
+      const faction = factions[owner];
       return faction ? { fill: `#${faction.primaryColour}` } : null;
-    },
-    tooltipRegionOwner(region) {
-      const owner = this.factionList[region.key];
-      const faction = this.factions[owner];
+    };
+
+    const tooltipRegionOwner = (region) => {
+      const owner = ownedRegions.value[region.key];
+      const faction = factions[owner];
 
       return {
         type: "region-owner",
         key: region.key,
         owner: faction ? faction.key : null
       };
-    }
+    };
+
+    const paintRegion = (region) => {
+      if (selectedFaction.value) {
+        ownedRegions.value[region.key] = selectedFaction.value.key;
+      } else {
+        ownedRegions.value[region.key] = null;
+      }
+    };
+
+    return {
+      regions,
+      style,
+      tooltipRegionOwner,
+      paintRegion
+    };
   }
 };
 </script>