+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 2019-12-17
+## 2019-12-19
+### Added
+- Shareable link for planner/painter
+### Fixed
+- Missing faction flags
+- Missing factions in faction selector
+
+## 2019-12-16
 ### Added
 - Changelog
 - Improved update dialog